# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eBPFsentinel Agent API",
    "description": "REST API for the eBPFsentinel network security agent.\n\nProvides endpoints for managing firewall rules, IPS/IDS policies, rate limiting, L7 filtering, threat intelligence, alerts, and audit logs.\n\nMost endpoints require authentication via JWT Bearer token or API key.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "v0.0.0-dev"
  },
  "paths": {
...
    "/api/v1/domains/blocklist": {
      "post": {
        "tags": [
          "Domain Intelligence"
        ],
        "summary": "`POST /api/v1/domains/blocklist` - add a domain to the runtime blocklist.",
        "operationId": "add_to_blocklist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlocklistAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Domain added to blocklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlocklistAddResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain pattern",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "DNS blocklist not enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/domains/blocklist/{domain}": {
      "delete": {
        "tags": [
          "Domain Intelligence"
        ],
        "summary": "`DELETE /api/v1/domains/blocklist/{domain}` - remove a domain from the runtime blocklist.",
        "operationId": "remove_from_blocklist",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "description": "Domain pattern to remove",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain removed from blocklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlocklistRemoveResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found in blocklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "DNS blocklist not enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/domains/reputation": {
      "get": {
        "tags": [
          "Domain Intelligence"
        ],
        "summary": "`GET /api/v1/domains/reputation` - query domain reputations.",
        "operationId": "list_domain_reputations",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "description": "Filter by domain (exact match).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "description": "Minimum reputation score (0.0-1.0).",
            "required": false,
            "schema": {
              "type": [
                "number",
                "null"
              ],
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-indexed).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Page size (max 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain reputations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainReputationListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Domain reputation not enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlocklistAddRequest": {
        "type": "object",
        "required": [
          "domain"
        ],
        "properties": {
          "domain": {
            "type": "string"
          }
        }
      },
      "BlocklistAddResponse": {
        "type": "object",
        "required": [
          "domain",
          "added"
        ],
        "properties": {
          "added": {
            "type": "boolean"
          },
          "domain": {
            "type": "string"
          }
        }
      },
      "BlocklistRemoveResponse": {
        "type": "object",
        "required": [
          "domain",
          "removed"
        ],
        "properties": {
          "domain": {
            "type": "string"
          },
          "removed": {
            "type": "boolean"
          }
        }
      },
...
      "DomainReputationListResponse": {
        "type": "object",
        "required": [
          "entries",
          "page",
          "page_size"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DomainReputationResponse"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "DomainReputationResponse": {
        "type": "object",
        "required": [
          "domain",
          "score",
          "factors",
          "first_seen",
          "last_seen",
          "is_blocked"
        ],
        "properties": {
          "domain": {
            "type": "string"
          },
          "factors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "first_seen": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "is_blocked": {
            "type": "boolean"
          },
          "last_seen": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "score": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `RULE_NOT_FOUND`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key configured in agent"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained via OIDC or static configuration"
      }
    }
  },
...
}